Restructure dags and configs by team
