+++ configs/solutions_team/tensorflow/common.py
