+++ dags/examples/xpk_example_dag.py
@@ -17,7 +17,7 @@
 import datetime
 from airflow import models
 from configs.vm_resource import TpuVersion, Project, Zone, ClusterName, DockerImage
-from configs.example import xpk_example_config as config
+from configs.examples import xpk_example_config as config
 
 
 # TODO(ranran): add following examples:
