+++ dags/solutions_team/solutionsteam_pax_latest_supported.py
@@ -18,7 +18,7 @@ import datetime
 from airflow import models
 from configs import composer_env, gcs_bucket
 from configs.vm_resource import TpuVersion, Zone
-from configs.xlml.pax import solutionsteam_pax_supported_config as pax_config
+from configs.solutions_team.pax import solutionsteam_pax_supported_config as pax_config
 
 
 # Run once a day at 10 am UTC (2 am PST)
